pyproject.toml
[project]
name = "theresa"
version = "1.0.0"
description = "qq-bot framework"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.3",
    "asyncpg>=0.31.0",
    "colorama==0.4.6",
    "emoji<2.0.0",
    "fastapi>=0.118.0",
    "uvicorn>=0.37.0",
    "httpx[socks]>=0.28.1",
    "jinja2>=3.1.6",
    "openai==2.11.0",
    "pillow==12.1.1",
    "pilmoji==2.0.5",
    "redis==7.1.0",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.46",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pre-commit-uv>=4.1.5",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.0",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
